Day 11 - Part 2 was too difficult to implement in C

+++ data-structures/node.h
@@ -5,7 +5,7 @@
 
 typedef struct Node 
 {
-    int val;
+    long long val;
     struct Node* next;
     struct Node* prev;
 } Node;
